+++ service/interceptor/paramcheck/check.go
@@ -25,13 +25,7 @@ import (
 	"github.com/polarismesh/specification/source/go/api/v1/traffic_manage"
 
 	"github.com/polarismesh/polaris/common/api/l5"
-	"github.com/polarismesh/polaris/common/utils"
 
-	apifault "github.com/polarismesh/specification/source/go/api/v1/fault_tolerance"
-	apimodel "github.com/polarismesh/specification/source/go/api/v1/model"
-	apiservice "github.com/polarismesh/specification/source/go/api/v1/service_manage"
-
-	api "github.com/polarismesh/polaris/common/api/v1"
 	"github.com/polarismesh/polaris/common/model"
 )
 
@@ -40,30 +34,6 @@ func (svr *Server) AppendServiceContractInterfaces(ctx context.Context,
 	contract *service_manage.ServiceContract,
 	source service_manage.InterfaceDescriptor_Source) *service_manage.Response {
 	return svr.nextSvr.AppendServiceContractInterfaces(ctx, contract, source)
-}
-
-// CreateCircuitBreakerRules implements service.DiscoverServer.
-func (svr *Server) CreateCircuitBreakerRules(ctx context.Context,
-	request []*fault_tolerance.CircuitBreakerRule) *service_manage.BatchWriteResponse {
-	return svr.nextSvr.CreateCircuitBreakerRules(ctx, request)
-}
-
-// CreateCircuitBreakerVersions implements service.DiscoverServer.
-func (svr *Server) CreateCircuitBreakerVersions(ctx context.Context,
-	req []*fault_tolerance.CircuitBreaker) *service_manage.BatchWriteResponse {
-	return svr.nextSvr.CreateCircuitBreakerVersions(ctx, req)
-}
-
-// CreateCircuitBreakers implements service.DiscoverServer.
-func (svr *Server) CreateCircuitBreakers(ctx context.Context,
-	req []*fault_tolerance.CircuitBreaker) *service_manage.BatchWriteResponse {
-	return svr.nextSvr.CreateCircuitBreakers(ctx, req)
-}
-
-// CreateFaultDetectRules implements service.DiscoverServer.
-func (svr *Server) CreateFaultDetectRules(ctx context.Context,
-	request []*fault_tolerance.FaultDetectRule) *service_manage.BatchWriteResponse {
-	return svr.nextSvr.CreateFaultDetectRules(ctx, request)
 }
 
 // CreateInstances implements service.DiscoverServer.
@@ -203,12 +173,6 @@ func (svr *Server) GetInstancesCount(ctx context.Context) *service_manage.BatchQ
 	return svr.nextSvr.GetInstancesCount(ctx)
 }
 
-// GetMasterCircuitBreakers implements service.DiscoverServer.
-func (svr *Server) GetMasterCircuitBreakers(ctx context.Context,
-	query map[string]string) *service_manage.BatchQueryResponse {
-	return svr.nextSvr.GetMasterCircuitBreakers(ctx, query)
-}
-
 // GetPrometheusTargets implements service.DiscoverServer.
 func (svr *Server) GetPrometheusTargets(ctx context.Context,
 	query map[string]string) *model.PrometheusDiscoveryResponse {
@@ -221,70 +185,10 @@ func (svr *Server) GetRateLimits(ctx context.Context,
 	return svr.nextSvr.GetRateLimits(ctx, query)
 }
 
-// GetReleaseCircuitBreakers implements service.DiscoverServer.
-func (svr *Server) GetReleaseCircuitBreakers(ctx context.Context,
-	query map[string]string) *service_manage.BatchQueryResponse {
-	return svr.nextSvr.GetReleaseCircuitBreakers(ctx, query)
-}
-
-// GetCircuitBreaker implements service.DiscoverServer.
-func (svr *Server) GetCircuitBreaker(ctx context.Context,
-	query map[string]string) *service_manage.BatchQueryResponse {
-	return svr.nextSvr.GetCircuitBreaker(ctx, query)
-}
-
-// GetCircuitBreakerByService implements service.DiscoverServer.
-func (svr *Server) GetCircuitBreakerByService(ctx context.Context,
-	query map[string]string) *service_manage.BatchQueryResponse {
-	return svr.nextSvr.GetCircuitBreakerByService(ctx, query)
-}
-
-// DeleteCircuitBreakers implements service.DiscoverServer.
-func (svr *Server) DeleteCircuitBreakers(ctx context.Context,
-	req []*fault_tolerance.CircuitBreaker) *service_manage.BatchWriteResponse {
-	return svr.nextSvr.DeleteCircuitBreakers(ctx, req)
-}
-
-// GetCircuitBreakerToken implements service.DiscoverServer.
-func (svr *Server) GetCircuitBreakerToken(ctx context.Context,
-	req *fault_tolerance.CircuitBreaker) *service_manage.Response {
-	return svr.nextSvr.GetCircuitBreakerToken(ctx, req)
-}
-
-// GetCircuitBreakerVersions implements service.DiscoverServer.
-func (svr *Server) GetCircuitBreakerVersions(ctx context.Context,
-	query map[string]string) *service_manage.BatchQueryResponse {
-	return svr.nextSvr.GetCircuitBreakerVersions(ctx, query)
-}
-
-// GetCircuitBreakerRules implements service.DiscoverServer.
-func (svr *Server) GetCircuitBreakerRules(ctx context.Context,
-	query map[string]string) *service_manage.BatchQueryResponse {
-	return svr.nextSvr.GetCircuitBreakerRules(ctx, query)
-}
-
-// DeleteCircuitBreakerRules implements service.DiscoverServer.
-func (svr *Server) DeleteCircuitBreakerRules(ctx context.Context,
-	request []*fault_tolerance.CircuitBreakerRule) *service_manage.BatchWriteResponse {
-	if err := checkBatchCircuitBreakerRules(request); err != nil {
-		return err
-	}
-	return svr.nextSvr.DeleteCircuitBreakerRules(ctx, request)
-}
-
 // DeleteFaultDetectRules implements service.DiscoverServer.
 func (svr *Server) DeleteFaultDetectRules(ctx context.Context,
 	request []*fault_tolerance.FaultDetectRule) *service_manage.BatchWriteResponse {
 	return svr.nextSvr.DeleteFaultDetectRules(ctx, request)
-}
-
-// EnableCircuitBreakerRules implements service.DiscoverServer.
-func (svr *Server) EnableCircuitBreakerRules(ctx context.Context,
-	request []*fault_tolerance.CircuitBreakerRule) *service_manage.BatchWriteResponse {
-	if err := checkBatchCircuitBreakerRules(request); err != nil {
-		return err
-	}
-	return svr.nextSvr.EnableCircuitBreakerRules(ctx, request)
 }
 
 // GetFaultDetectRules implements service.DiscoverServer.
@@ -353,27 +257,10 @@ func (svr *Server) SyncByAgentCmd(ctx context.Context, sbac *l5.Cl5SyncByAgentCm
 	return svr.nextSvr.SyncByAgentCmd(ctx, sbac)
 }
 
-// ReleaseCircuitBreakers implements service.DiscoverServer.
-func (svr *Server) ReleaseCircuitBreakers(ctx context.Context, req []*service_manage.ConfigRelease) *service_manage.BatchWriteResponse {
-	return svr.nextSvr.ReleaseCircuitBreakers(ctx, req)
-}
-
-// UnBindCircuitBreakers implements service.DiscoverServer.
-func (svr *Server) UnBindCircuitBreakers(ctx context.Context, req []*service_manage.ConfigRelease) *service_manage.BatchWriteResponse {
-	return svr.nextSvr.UnBindCircuitBreakers(ctx, req)
-}
-
-// UpdateCircuitBreakerRules implements service.DiscoverServer.
-func (svr *Server) UpdateCircuitBreakerRules(ctx context.Context, request []*fault_tolerance.CircuitBreakerRule) *service_manage.BatchWriteResponse {
-	if err := checkBatchCircuitBreakerRules(request); err != nil {
-		return err
-	}
-	return svr.nextSvr.UpdateCircuitBreakerRules(ctx, request)
-}
-
-// UpdateCircuitBreakers implements service.DiscoverServer.
-func (svr *Server) UpdateCircuitBreakers(ctx context.Context, req []*fault_tolerance.CircuitBreaker) *service_manage.BatchWriteResponse {
-	return svr.nextSvr.UpdateCircuitBreakers(ctx, req)
+// CreateFaultDetectRules implements service.DiscoverServer.
+func (svr *Server) CreateFaultDetectRules(ctx context.Context,
+	request []*fault_tolerance.FaultDetectRule) *service_manage.BatchWriteResponse {
+	return svr.nextSvr.CreateFaultDetectRules(ctx, request)
 }
 
 // UpdateFaultDetectRules implements service.DiscoverServer.
@@ -419,15 +306,4 @@ func (svr *Server) UpdateServiceToken(ctx context.Context, req *service_manage.S
 // UpdateServices implements service.DiscoverServer.
 func (svr *Server) UpdateServices(ctx context.Context, req []*service_manage.Service) *service_manage.BatchWriteResponse {
 	return svr.nextSvr.UpdateServices(ctx, req)
-}
-
-func checkBatchCircuitBreakerRules(req []*apifault.CircuitBreakerRule) *apiservice.BatchWriteResponse {
-	if len(req) == 0 {
-		return api.NewBatchWriteResponse(apimodel.Code_EmptyRequest)
-	}
-
-	if len(req) > utils.MaxBatchSize {
-		return api.NewBatchWriteResponse(apimodel.Code_BatchSizeOverLimit)
-	}
-	return nil
 }
